+++ lib/use_cases/slack/send_messages.rb
@@ -4,7 +4,7 @@ module UseCases
       def initialize(
         slack_gateway: Gateways::SlackMessage.new,
         team_usecase: UseCases::FetchTeams.new,
-        pull_request_usecase: UseCases::FetchPullRequests.new,
+        pull_request_usecase: UseCases::PullRequests::Fetch.new,
         group_applications_by_team_usecase: UseCases::Group::ApplicationsByTeam.new,
         scheduler: UseCases::Slack::Schedulers::EveryDay.new,
         message_presenter:
